added type.c file and type structure
