+++ CS_header.h
@@ -35,6 +35,7 @@
 	
 // CSpace's functions sorted by dependance (so DON'T MODIFY the order if you don't want to fight dozen of gcc's error)
 	#include "string functions.c"
+	#include "type.c"
 	#include "system functions.c"
 	#include "mathematical functions.c"
 	#include "OnlyPrintfSystem.c"
